StuntGuard-ML/requirement.txt
Flask==2.0.2
numpy==1.21.3
mysql-connector-python==8.0.28
python-dotenv==0.19.2
tensorflow==2.7.0